Code cleanup, Fixed paging scroll to top, testing extension

+++ app/src/main/java/com/fanulhaq/githubuser/data/repository/detail/DetailRepoImpl.kt
@@ -40,8 +40,8 @@ class DetailRepoImpl @Inject constructor(
                     if(!company.isNullOrEmpty()) bio = "$company ${
                         if(!bio.isNullOrEmpty()) ", $bio" else ""
                     }"
-                    followers = followers?.toInt()?.numberShortFormatter()
-                    following = following?.toInt()?.numberShortFormatter()
+                    followers = followers?.toLong()?.numberShortFormatter()
+                    following = following?.toLong()?.numberShortFormatter()
                 }
                 detailDao.insert(response)
             }
